스토리보드 전체 수정: 네비게이션 컨트롤러 위치 변경
