+++ DOTORI/Controller/MyPageViewController.swift
@@ -13,8 +13,9 @@ class MyPageViewController: UIViewController, WKNavigationDelegate {
     var myPostings: [PostingInfo] = []
     let webView = WKWebView()
     var selectedUserName : String? //디테일페이지에서 클릭한 프로필의 유저 이름
-    @IBOutlet weak var mainTitle: UILabel!
-    @IBOutlet weak var mySetting: UIButton!
+    
+    @IBOutlet weak var titleButton: UIBarButtonItem!
+    @IBOutlet weak var mySetting: UIBarButtonItem!
     @IBOutlet weak var profileImage: UIImageView!
     @IBOutlet weak var name: UILabel!
     @IBOutlet weak var nickname: UILabel!
@@ -25,11 +26,12 @@ class MyPageViewController: UIViewController, WKNavigationDelegate {
     @IBOutlet weak var blogUrl: UIButton!
     @IBOutlet weak var github: UILabel!
     @IBOutlet weak var githubUrl: UIButton!
-    
     @IBOutlet weak var tableView: UITableView!
     
-    @IBAction func mySetting(_ sender: UIButton) {
+    @IBAction func mySetting(_ sender: UIBarButtonItem) {
+        print("클릭")
     }
+    
     
     @IBAction func blogUrl(_ sender: UIButton) {
         if let urlText = blogUrl.titleLabel?.text {
@@ -50,11 +52,9 @@ class MyPageViewController: UIViewController, WKNavigationDelegate {
     
     override func viewDidLoad() {
         super.viewDidLoad()
-        if let text = selectedUserName
-        {
-            name.text = text
-        }
         
+        
+        if let text = selectedUserName { name.text = text }
         loadAccount()
         tableView.dataSource = self
         tableView.delegate = self
@@ -115,46 +115,6 @@ extension MyPageViewController: UITableViewDataSource, UITableViewDelegate {
     }
 }
 
-
-class MyPostingTableViewCell: UITableViewCell, UITextViewDelegate {
-    @IBOutlet weak var profileImage: UIImageView!
-    @IBOutlet weak var name: UILabel!
-    @IBOutlet weak var nickname: UILabel!
-    @IBOutlet weak var createTime: UILabel!
-    @IBOutlet weak var content: UITextView!
-    @IBOutlet weak var contentImage: UIImageView!
-    @IBOutlet weak var postingSetting: UIButton!
-    
-    func setupUI(posting: PostingInfo) {
-        profileImage.image = posting.user.profileImage ?? UIImage(named: "defaultProfileImage")
-        profileImage.layer.cornerRadius = profileImage.frame.size.width / 2
-        profileImage.clipsToBounds = true
-        name.text = posting.user.name
-        nickname.text = posting.user.nickname
-        createTime.text = posting.createTime.GetCurrentTime()
-        content.text = posting.content
-        contentImage.image = posting.contentImage ?? UIImage(named: "defaultProfileImage")
-        contentImage.layer.cornerRadius = 10
-        contentImage.clipsToBounds = true
-
-    }
-    
-    override func awakeFromNib() {
-        super.awakeFromNib()
-            
-        print("버튼 클릭: postingSetting")
-        let seletedpostingSetting = {(action: UIAction) in
-            }
-        postingSetting.menu = UIMenu(children: [
-            UIAction(title: "게시물 삭제", image: UIImage(systemName: "trash"),attributes: .destructive, handler: seletedpostingSetting)])
-        postingSetting.showsMenuAsPrimaryAction = true
-        postingSetting.changesSelectionAsPrimaryAction = false
-        content.delegate = self
-        content.isScrollEnabled = false
-        content.sizeToFit()
-    }
-    
-}
 
 class WebViewController: UIViewController {
     @IBOutlet weak var webView: WKWebView!
