+++ DOTORI/View/DOTORI/View/Base.lproj/SearchViewController.storyboard
@@ -9,6 +9,22 @@
         <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
     </dependencies>
     <scenes>
+        <!--Search Navigation Controller-->
+        <scene sceneID="LJV-pE-mHX">
+            <objects>
+                <navigationController storyboardIdentifier="SearchNavigationController" id="V0x-Fg-LzN" customClass="SearchNavigationController" sceneMemberID="viewController">
+                    <navigationBar key="navigationBar" contentMode="scaleToFill" id="BI8-GR-9QX">
+                        <rect key="frame" x="0.0" y="59" width="393" height="44"/>
+                        <autoresizingMask key="autoresizingMask"/>
+                    </navigationBar>
+                    <connections>
+                        <segue destination="VwU-Ww-U3g" kind="relationship" relationship="rootViewController" id="vkq-Sc-XOC"/>
+                    </connections>
+                </navigationController>
+                <placeholder placeholderIdentifier="IBFirstResponder" id="9c2-0K-4r7" userLabel="First Responder" customClass="UIResponder" sceneMemberID="firstResponder"/>
+            </objects>
+            <point key="canvasLocation" x="-331" y="213"/>
+        </scene>
         <!--Search View Controller-->
         <scene sceneID="VCQ-PX-92g">
             <objects>
@@ -16,20 +32,12 @@
                     <view key="view" contentMode="scaleToFill" id="fdi-QO-YcM">
                         <rect key="frame" x="0.0" y="0.0" width="393" height="852"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
-                        <subviews>
-                            <navigationBar contentMode="scaleToFill" fixedFrame="YES" translatesAutoresizingMaskIntoConstraints="NO" id="TMx-sR-w5I">
-                                <rect key="frame" x="0.0" y="59" width="393" height="44"/>
-                                <autoresizingMask key="autoresizingMask" widthSizable="YES" flexibleMaxY="YES"/>
-                                <items>
-                                    <navigationItem title="Title" id="yIu-au-h6p"/>
-                                </items>
-                            </navigationBar>
-                        </subviews>
                         <viewLayoutGuide key="safeArea" id="U5p-f6-hEY"/>
                         <color key="backgroundColor" systemColor="systemBackgroundColor"/>
                     </view>
                     <tabBarItem key="tabBarItem" title="" image="magnifyingglass" catalog="system" id="slB-oj-l92"/>
                     <toolbarItems/>
+                    <navigationItem key="navigationItem" id="hM7-eg-tHc"/>
                     <simulatedTabBarMetrics key="simulatedBottomBarMetrics"/>
                 </viewController>
                 <placeholder placeholderIdentifier="IBFirstResponder" id="i0D-z3-pVX" userLabel="First Responder" customClass="UIResponder" sceneMemberID="firstResponder"/>
